add `authorize` HOF
